+++ code/DPM.Applications/Features/Ships/UpdateShipLocationById/UpdateShipLocationByIdCommand.cs
@@ -10,6 +10,6 @@ namespace DPM.Applications.Features.Ships.UpdateShipLocationById
     public class UpdateShipLocationByIdCommand: IRequest<bool>
     {
         public long Id { get; set; }
-        public long[] Position { get; set; }
+        public double[] Position { get; set; }
     }
 }
